Don't remove the proposal node before proposing
Previously, we were first removing the proposal node and then
substituting the proposed programs into the hole left by the removal.
This was causing issues, because the removal sometimes resulted in
simplification of the resulting hole, meaning the proposal spot
was no longer a hole, but a concrete `RuleNode`. Instead, we can
directly substitute the proposal programs into their spot in the
tree that was chosen for proposing.

Does the word proposal sound weird yet...?